Move ContrastiveDataLoader and ContrastiveDataSplitter to external
