+++ scvi/dataloaders/__init__.py
@@ -3,9 +3,7 @@ from scvi.data import AnnTorchDataset
 
 from ._ann_dataloader import AnnDataLoader
 from ._concat_dataloader import ConcatDataLoader
-from ._contrastive_dataloader import ContrastiveDataLoader
 from ._data_splitting import (
-    ContrastiveDataSplitter,
     DataSplitter,
     DeviceBackedDataSplitter,
     SemiSupervisedDataSplitter,
@@ -21,6 +19,4 @@ __all__ = [
     "SemiSupervisedDataLoader",
     "DataSplitter",
     "SemiSupervisedDataSplitter",
-    "ContrastiveDataLoader",
-    "ContrastiveDataSplitter",
 ]
