+++ scvi/external/contrastivevi/__init__.py
@@ -1,4 +1,11 @@
+from ._contrastive_data_splitting import ContrastiveDataSplitter
+from ._contrastive_dataloader import ContrastiveDataLoader
 from ._model import ContrastiveVI
 from ._module import ContrastiveVAE
 
-__all__ = ["ContrastiveVI", "ContrastiveVAE"]
+__all__ = [
+    "ContrastiveDataLoader",
+    "ContrastiveDataSplitter",
+    "ContrastiveVAE",
+    "ContrastiveVI",
+]
